Refactor MutableAttributes.
