+++ SplunkAgent/Sources/SplunkAgent/Agent/Attributes/MutableAttributes+Codable.swift
@@ -1,0 +1,48 @@
+//
+/*
+Copyright 2025 Splunk Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+import Foundation
+
+extension MutableAttributes: Codable {
+
+    // MARK: - Codable
+
+    public func encode(to encoder: Encoder) throws {
+        var container = encoder.container(keyedBy: StringCodingKey.self)
+        let dictionary = attributes.getAll()
+
+        for (key, value) in dictionary {
+            try container.encode(value, forKey: StringCodingKey(stringValue: key)!)
+        }
+    }
+
+    // Helper for coding with string keys
+    struct StringCodingKey: CodingKey {
+        var stringValue: String
+        var intValue: Int?
+
+        init?(stringValue: String) {
+            self.stringValue = stringValue
+            intValue = nil
+        }
+
+        init?(intValue: Int) {
+            stringValue = String(intValue)
+            self.intValue = intValue
+        }
+    }
+}
