+++ SplunkAgent/Sources/SplunkAgent/Agent/Attributes/MutableAttributes+Equatable.swift
@@ -1,0 +1,46 @@
+//
+/*
+Copyright 2025 Splunk Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+import Foundation
+
+extension MutableAttributes: Equatable {
+
+    // MARK: - Equatable
+
+    public static func == (lhs: MutableAttributes, rhs: MutableAttributes) -> Bool {
+        let lhsDict = lhs.attributes.getAll()
+        let rhsDict = rhs.attributes.getAll()
+
+        // Compare dictionary sizes first for quick check
+        guard lhsDict.count == rhsDict.count else {
+            return false
+        }
+
+        // Compare each key-value pair
+        for (key, lhsValue) in lhsDict {
+            guard let rhsValue = rhsDict[key] else {
+                return false
+            }
+
+            if lhsValue != rhsValue {
+                return false
+            }
+        }
+
+        return true
+    }
+}
