+++ SplunkCommon/Sources/SplunkCommon/PackageIdentification/PackageIdentifier+InstanceName.swift
@@ -14,7 +14,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-	
+
 public extension PackageIdentifier {
 
     // MARK: - Package identification for instances
